hw-6-email
Updated user model with varify and varificationToken. Used Joi schema for it.
Created verify.js and resendEmail in controllers. for resending email and verify for verification. Also updated register logic for user with adding varificationToken upon registration and sending this verify mail to users email. Added new routes in auth for get and post.